# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 智慧云脑诊疗平台 API
  version: 0.1.0
  description: |
    阶段二：基础数据管理（认证 + 科室 + 患者 + 医生 + 药品）
    阶段三：排班与挂号（Phase 5-6）
    阶段四：分诊与就诊（Phase 7-8）

    当前包含接口：
    - 认证模块：login / logout / change-password（3 个）
    - 科室模块：tree / list / create / update（4 个）
    - 患者模块：register / me / detail / update / profile / search（6 个）
    - 医生模块：list / create / update / by-department（4 个）
    - 药品模块：list / detail / by-code / by-category / interaction / contraindications（7 个）
    - 排班模块：create / update / cancel / detail / by-doctor / by-department / available（7 个）
    - 挂号模块：create / cancel / detail / by-patient / by-doctor / pending / all（7 个）
    - 分诊模块：analyze / detail / by-patient（3 个）
    - 就诊模块：start / wait-exam / resume / complete / cancel / diagnoses / detail（12 个）

    契约来源：contracts/30_接口数据与错误契约.md 第11节 + 第12节
servers:
  - url: http://localhost:8080
    description: 本地后端
  - url: http://localhost:4010
    description: Prism API Mock Server
paths:
...
  /api/encounters/start:
    post:
      tags:
        - 就诊
      summary: 开始接诊（CREATED → IN_PROGRESS）
      description: |
        医生开始接诊患者。创建 Encounter 并同步 Appointment 状态为 IN_PROGRESS。
        一个 Appointment 只能关联一个有效 Encounter。
      operationId: startEncounter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/triage_encounter.yaml#/EncounterStartRequest'
      responses:
        '200':
          description: 开始接诊成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '409':
          description: 重复就诊或挂号状态无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/wait-exam:
    post:
      tags:
        - 就诊
      summary: 等待检查（IN_PROGRESS → WAITING_EXAM）
      operationId: waitForExam
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 状态更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/resume:
    post:
      tags:
        - 就诊
      summary: 继续诊疗（WAITING_EXAM → IN_PROGRESS）
      operationId: resumeEncounter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 继续诊疗成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '409':
          description: 状态冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/complete:
    post:
      tags:
        - 就诊
      summary: 完成就诊（IN_PROGRESS → COMPLETED）
      description: |
        完成就诊前置条件检查：
        - 病历已确认
        - 存在医生最终诊断
        - 所有检查检验已完成/已审核
        - 处方状态为 CONFIRMED 或 VOIDED（如有）
      operationId: completeEncounter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 完成就诊成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '409':
          description: 前置条件不满足（缺少诊断/病历未确认/检查未完成/处方状态异常）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/cancel:
    post:
      tags:
        - 就诊
      summary: 取消就诊（CREATED → CANCELLED）
      description: 仅 CREATED 状态可取消。IN_PROGRESS、WAITING_EXAM、COMPLETED 不允许取消。
      operationId: cancelEncounter
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: './schemas/triage_encounter.yaml#/EncounterCancelRequest'
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '409':
          description: 状态不允许取消
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}:
    get:
      tags:
        - 就诊
      summary: 获取就诊详情
      operationId: getEncounterById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 就诊详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '404':
          description: 就诊不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/appointment/{appointmentId}:
    get:
      tags:
        - 就诊
      summary: 按挂号 ID 查询就诊
      operationId: getEncounterByAppointmentId
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 就诊信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterResponse'
        '404':
          description: 就诊不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/doctor/{doctorId}:
    get:
      tags:
        - 就诊
      summary: 查询医生就诊列表（分页）
      operationId: getEncountersByDoctor
      security:
        - bearerAuth: []
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 就诊列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/patient/{patientId}:
    get:
      tags:
        - 就诊
      summary: 查询患者就诊列表（分页）
      operationId: getEncountersByPatient
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 就诊列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/diagnoses/ai:
    post:
      tags:
        - 就诊
      summary: 添加 AI 候选诊断
      description: |
        诊断隔离原则：AI 只能创建 type=PRELIMINARY、source=AI_SUGGESTION 的诊断。
        AI 不得创建 FINAL+DOCTOR 记录。
      operationId: addAIDiagnosis
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/triage_encounter.yaml#/EncounterDiagnosisRequest'
      responses:
        '200':
          description: AI 诊断添加成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterDiagnosisResponse'
        '409':
          description: 诊断隔离违规
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/diagnoses/doctor:
    post:
      tags:
        - 就诊
      summary: 添加医生最终诊断
      description: |
        诊断隔离原则：医生诊断必须为 type=FINAL、source=DOCTOR。
        医生不得创建非 FINAL 的诊断（类型冲突）。
      operationId: addDoctorDiagnosis
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/triage_encounter.yaml#/EncounterDiagnosisRequest'
      responses:
        '200':
          description: 医生诊断添加成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: './schemas/triage_encounter.yaml#/EncounterDiagnosisResponse'
        '409':
          description: 诊断类型冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/encounters/{id}/diagnoses:
    get:
      tags:
        - 就诊
      summary: 查询就诊诊断列表
      operationId: getEncounterDiagnoses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 诊断列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: './schemas/triage_encounter.yaml#/EncounterDiagnosisResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      $ref: ./schemas/common.yaml#/ApiResponse
    PageResponse:
      $ref: ./schemas/common.yaml#/PageResponse
...